set everything to float to avoid conversions

+++ splineDrive.cpp
@@ -39,18 +39,18 @@ void setCurve(Distortion *distortion, float gain, int preserveDynamics)
    distortion->oldPreserveDynamics = preserveDynamics;
    
    // NOTE: decent drive
-   //std::vector<double> x = {0.0, 0.1, 1.0};
-   std::vector<double> x = {-1.0, -0.1, 0.0, 0.1, 1.0};
-   //std::vector<double> y = {0.0, min(0.1 * gain, 1.0), 1.0};
-   std::vector<double> y = {-1.0, 0.0 - min(0.1 * gain, 1.0), 0.0, min(0.1 * gain, 1.0), 1.0};
+   //std::vector<float> x = {0.0, 0.1, 1.0};
+   std::vector<float> x = {-1.0f, -0.1f, 0.0f, 0.1f, 1.0f};
+   //std::vector<float> y = {0.0, min(0.1 * gain, 1.0), 1.0};
+   std::vector<float> y = {-1.0f, 0.0f - min(0.1f * gain, 1.0f), 0.0f, min(0.1f * gain, 1.0f), 1.0f};
 
    // NOTE: Crazy curve, noisy when plaid loud
-   // std::vector<double> x = {0.0, 0.1, 0.2, 0.5, 1.0};
-   // std::vector<double> y = {0.0, 0.8, -0.8, 0.7, 1.0};
+   // std::vector<float> x = {0.0f, 0.1f, 0.2f, 0.5f, 1.0f};
+   // std::vector<float> y = {0.0f, 0.8f, -0.8f, 0.7f, 1.0f};
 
    // NOTE: Kind of octavey still noisy
-   // std::vector<double> x = {0.0, 0.1, 0.2, 1.0};
-   // std::vector<double> y = {0.0, -0.9, 0.8, 1.0};
+   // std::vector<float> x = {0.0f, 0.1f, 0.2f, 1.0f};
+   // std::vector<float> y = {0.0f, -0.9f, 0.8f, 1.0f};
 
    distortion->spline.set_points(x, y, tk::spline::cspline);
    
@@ -61,7 +61,6 @@ void setCurve(Distortion *distortion, float gain, int preserveDynamics)
    pFile = fopen("splineDrive.log", "a+");
    fprintf(pFile, "gain %.2f\n", gain);
    fprintf(pFile, "preserveDynamis: %s\n", preserveDynamics > 0 ? "true" : "false");
-
 
    // for (float i = -1.0; i <= 1.01; i+=0.05)
    // {
